+++ src/arch/toy.rs
@@ -312,17 +312,17 @@ fn decode_operands(desc: &InstDesc, va: VA, img: &[u8], ops: &mut [Operand; 2])
 			(2, None)
 		}
 		S8 => {
-			let target = VA(((va.0 as isize) + (img[1] as i8 as isize) + 2) as usize);
+			let target = VA(((va.0 as isize) + (img[1] as i8 as isize) + 2) as u64);
 			ops[0] = Operand::Mem(target, MemAccess::Target);
 			(1, Some(target))
 		}
 		I16 => {
-			let target = VA(u16::from_le_bytes(img[1..].try_into().unwrap()) as usize);
+			let target = VA(u16::from_le_bytes(img[1..].try_into().unwrap()) as u64);
 			ops[0] = Operand::Mem(target, MemAccess::Target);
 			(1, Some(target))
 		}
 		RI16 => {
-			let addr = VA(u16::from_le_bytes(img[1..].try_into().unwrap()) as usize);
+			let addr = VA(u16::from_le_bytes(img[1..].try_into().unwrap()) as u64);
 			ops[0] = opcode_reg;
 			ops[1] = Operand::Mem(addr, desc.meta_op.access().unwrap());
 			(2, None)
